+++ packages/shared/src/constants/index.ts
@@ -42,4 +42,4 @@ export const ZOOM_STEP = 0.1;
 
 export const DEFAULT_ADMIN_EMAIL = 'admin@tactihub.local';
 
-export const APP_VERSION = '1.8.3';
+export const APP_VERSION = '1.8.4';